chore(masonry): rename older src
