+++ modules/masonry/src.old/brick/design0/BrickStatement.ts
@@ -1,12 +1,12 @@
-import type { TBrickRenderPropsExpression, TColor, TCoords, TExtent } from '@/@types/brick';
-import { BrickModelExpression } from '../model';
+import type { TBrickRenderPropsStatement, TColor, TCoords, TExtent } from 'src.old/@types/brick';
+import { BrickModelStatement } from '../model';
 import { generatePath } from '../utils/path';
 
 /**
  * @class
- * Final class that defines an expression brick.
+ * Final class that defines a statement brick.
  */
-export default class BrickExpression extends BrickModelExpression {
+export default class BrickStatement extends BrickModelStatement {
     readonly _pathResults: ReturnType<typeof generatePath>;
 
     private _boundingBoxArgs: Record<string, TExtent> = {};
@@ -15,24 +15,27 @@ export default class BrickExpression extends BrickModelExpression {
         uuid: string;
         name: string;
         label: string;
-        glyph?: string;
+        glyph: string;
+        args: { id: string; label: string }[];
         colorBg: TColor;
         colorFg: TColor;
         colorBgHighlight: TColor;
         colorFgHighlight: TColor;
         outline: TColor;
-        args: { id: string; label: string }[];
+        scale: number;
+        connectAbove: boolean;
+        connectBelow: boolean;
     }) {
         super(params);
 
         this._pathResults = generatePath({
             hasNest: false,
-            hasNotchArg: true,
-            hasNotchInsTop: false,
-            hasNotchInsBot: false,
+            hasNotchArg: false,
+            hasNotchInsTop: params.connectAbove,
+            hasNotchInsBot: params.connectBelow,
             scale: this._scale,
             innerLengthX: 100,
-            argHeights: Array.from({ length: this._args.length }, () => 17),
+            argHeights: Array.from({ length: params.args.length }, () => 17),
         });
     }
 
@@ -43,11 +46,18 @@ export default class BrickExpression extends BrickModelExpression {
         };
     }
 
-    public get connPointsFixed(): Record<'argOutgoing', { extent: TExtent; coords: TCoords }> {
+    public get connPointsFixed(): Record<
+        'insTop' | 'insBottom',
+        { extent: TExtent; coords: TCoords }
+    > {
         return {
-            argOutgoing: {
-                extent: this._pathResults.bBoxNotchArg!.extent,
-                coords: this._pathResults.bBoxNotchArg!.coords,
+            insTop: {
+                extent: this._pathResults.bBoxNotchInsTop!.extent,
+                coords: this._pathResults.bBoxNotchInsTop!.coords,
+            },
+            insBottom: {
+                extent: this._pathResults.bBoxNotchInsBot!.extent,
+                coords: this._pathResults.bBoxNotchInsBot!.coords,
             },
         };
     }
@@ -65,7 +75,7 @@ export default class BrickExpression extends BrickModelExpression {
         return results;
     }
 
-    public get renderProps(): TBrickRenderPropsExpression {
+    public get renderProps(): TBrickRenderPropsStatement {
         return {
             path: this._pathResults.path,
             label: this._label,
@@ -81,6 +91,14 @@ export default class BrickExpression extends BrickModelExpression {
 
     public setBoundingBoxArg(id: string, extent: TExtent): void {
         this._boundingBoxArgs[id] = extent;
+    }
+
+    public setConnectAbove(connectAbove: boolean): void {
+        this._connectAbove = connectAbove;
+    }
+
+    public setConnectBelow(connectBelow: boolean): void {
+        this._connectBelow = connectBelow;
     }
 
     public setHighlighted(highlighted: boolean): void {
