+++ modules/masonry/tsconfig.json
@@ -8,7 +8,7 @@
         "baseUrl": ".",
         "paths": {
             "@/*": [
-                "./src/*"
+                "./src/*",
             ],
             "#/@types/*": [
                 "../../@types/*"
